WIP: stdlib fdk experiment
note: this is totally a toy and doesn't implement an fdk interface, this just
sets out to talk the fn http unix socket protocol to see if it's possible to
do without a bunch of libraries. if this is possible, it's not much of a
stretch to add an interface around all this, I'm just not good at python and
will let somebody else do that probably.

this uses stdlib methods to make a unix socket server and serve http requests
this took a few hours to throw together after reading the stdlib docs and
searching for examples. got it returning a 200 anyway, so I mean think this
thing is okay for our purposes honestly? we don't need a robust http server
for this thing at all.