asyncio working with h11
this uses asyncio unix server from stdlib instead of the sync unix server from
stdlib, and then uses h11 to parse the http parts out. this probably doesn't
handle timeouts which we may want to do at some level (we get a deadline from
the fdk protocol) and flow control, which we may not need to handle given what
we're doing here. anyway, this thing works in the most toy-ish sense, and had
to upgrade to 3.7 to do asyncio.run - I'm not good at this, that's easy to
fix I would assume (as we should support 3.6? intention wasn't to change
python version support, at least, to be clear).

notably, this dependency is really small, and doesn't add to image build times
or size very much for our base image (like on master). h11 also has no other
dependencies itself. on master, on my laptop, the build loop was longer and
images were much larger on master. also, on master, on my laptop i'd get the 5
second startup timeout. with this function I do not get the 5 second startup
timeout when using h11 and the asyncio socket. at least, this is looking like
a pretty plausible option; my laptop is a relatively similarly constrained
environment for various reasons (thanks McAfee) so this bodes well. the code
itself, as you can see, it's about 50 LOC to do this, which is not bad at all
- the h11 dep itself is about 1k LOC, which is about what we were looking at
for our own http parser - which is an option, but the latter needs to write a
ton of tests, which h11 has already done, so...

if this path looks okay (we should discuss and weigh), then maybe or hopefully
somebody with some real python skills can pick this task up from me to make
the fdk all nice and actually parse our headers, etc, etc.