attempt fixing content type issue (#83)
buggy headers look like:

```
{Fn-Http-Status=[200], Opc-Request-Id=[/01DBJASGQY1BT1AHGZJ000PC8D/01DBJASGQY1BT1AHGZJ000PC8E], Connection=[keep-alive], Fn-Call-Id=[01DBJASGRW1BT1AHGZJ000PC8F], Fn-Http-H-Content-Type=[application/json], Content-Length=[462], Date=[Thu, 23 May 2019 12:31:35 GMT], Content-Type=[None]}
```

but should have `Content-Type: [application/json]` and not
`Content-Type:[None]` nor `Fn-Http-H-Content-Type`

* removes content type from encap headers, which was just ripping content type
out of the headers bucket anyway - just leave in the headers bucket, since it
wasn't removing it from the header bucket for usage, it was a confusing
contract.
* actually use the gateway headers if Fn-Intent=httprequest ("the contract"),
we weren't checking this at all or stripping gateway headers when we need to
* only encap the headers for gateway responses instead of for both invoke and
gateway responses
* no more default of `Content-Type: text/plain`
* fix up test fixtures

we should probably have separate types of contexts like things were before
this, but we should maybe use inheritance - I tried this, but didn't get it
working, though now I think it was a separate issue why I didn't get it
working. anyway, this fixes the bug(s) in invoke. gateway headers were not
being used nor well tested anyway (well, because it was defaulting to them
confusingly), so we could patch later.