properly grab content type
we're not using a case insensitive dict so these end up overlapping. this now
grabs the right content type out of the user headers before creating the real
http headers, at least, but we should probably use a case insensitive dict for
these things as this has been a real pain. this also changes the default to be
"text/plain" instead of sending in Content-Type: None is the user does not set
this particular header. Now I see application/json being set properly without
None or another value in content type, so this ought to work.