fix gateway headers / headers casing (#88)
the http library we're using automatically lower cases headers, and we weren't
properly handling different cases when encapsulating certain headers such as
content type and the gateway headers. this changes most of the headers to be
turned into lower case to check them and lower cases them going out of the
response in order to test them as well. over the wire itself, fn via the go
http library will end up turning these into the canonical header casing, but
for consistency in python this seems ideal. there are tests to ensure these
are checked properly in various cases now.

closes #87