Fix "None-None" race condition with keep-alive timeout
Keep-alives are handled by a scheduled routine, which closes the
connection if the keep-alive has expired. The routine calculates the
keep-alive time remaining based on the last response time, but that's
not set until after the routine is scheduled, creating a race
condition. The fix is to set the last response time before scheduling
the keep-alive routine. Additionally the calculation is a bit more
robust now, so the scheduled routine doesn't need to deal with None.

The async code is largely imported from [sanic](https://github.com/huge-success/sanic),
modified and stripped-down to reduce cold start times.

Addresses issue #103