Format logs with call ID and one-line exceptions
Python exception backtraces were sent as multiple separate lines. This
is inconvenient for logging services like Papertrail and OCI Logging
which create a record for each line. This patch formats the backtrace
as a single line.

The Call ID is logged with all log lines. This makes it easier to
identify which log lines belong to which call.

The current time is no longer logged.  Log destinations such as
Papertrail and OCI Logging record and display the log event time, so
this information is redundant.